+++ djpcms/views/docview.py
@@ -10,7 +10,7 @@ from django.utils.safestring import mark_safe
 from unipath import FSPath as Path
 
 from djpcms.views.baseview import djpcmsview
-from djpcms.views.appsite import ApplicationBase
+from djpcms.views.appview import ApplicationBase